requirements.txt
ollama
openai
slack_bolt
langchain
langchain_community
langchain-ollama
langchain-chroma
langchain-experimental
python-dotenv
termcolor
langchain-ollama
pytest
flake8
black
slack-cli-hooks
py-trello
PyGithub
faiss-cpu
unstructured
langchain-unstructured
python-magic-bin
